requirements-dev.txt
-r requirements.txt
flake8==3.0.4
freezegun==0.3.8
ipdb==0.10.2
pytest-cov==2.3.1
pytest-sugar==0.8.0
pytest==3.0.2
requests_mock==1.3.0
